+++ RemoteWorkBundle.php
@@ -10,8 +10,15 @@
 namespace KimaiPlugin\RemoteWorkBundle;
 
 use App\Plugin\PluginInterface;
+use KimaiPlugin\RemoteWorkBundle\DependencyInjection\Compiler\IconsCompilerPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class RemoteWorkBundle extends Bundle implements PluginInterface
 {
+    public function build(ContainerBuilder $container): void
+    {
+        parent::build($container);
+        $container->addCompilerPass(new IconsCompilerPass());
+    }
 }